% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggbiplot_geoms.R
\name{geom_arrow}
\alias{geom_arrow}
\title{Title}
\usage{
geom_arrow(mapping, data, inherit.aes = FALSE, arrow_size = 0.5, ...)
}
\arguments{
\item{mapping}{zie ggplot}

\item{data}{The data to be displayed in this layer}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. For creating a biplot this should be FALSE.}

\item{arrow_size}{size of the arrow heads}

\item{...}{other arguments passed on to layer.}
}
\description{
Title
}
