% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggbiplot_geoms.R
\name{geom_arrowlabel}
\alias{geom_arrowlabel}
\title{Title}
\usage{
geom_arrowlabel(mapping, data, inherit.aes = FALSE, ...)
}
\arguments{
\item{mapping}{typical mapping construct for ggplots. Typical in the form of "aes(x=, y=, color=, ...)"}

\item{data}{The data to be displayed in this layer}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. For creating a biplot this should be FALSE.}

\item{...}{other arguments passed on to layer.}
}
\description{
Title
}
