% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggbiplot_geoms.R
\name{geom_corcircle}
\alias{geom_corcircle}
\title{Title}
\usage{
geom_corcircle(radius, resolution = 200, ...)
}
\arguments{
\item{radius}{numeric radius of the correlation circle.}

\item{resolution}{how many points to use to draw the circle}

\item{...}{other arguments passed on to layer.}
}
\description{
Title
}
