% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggbiplot_helper.R
\name{geom_selector}
\alias{geom_selector}
\title{Title}
\usage{
geom_selector(mapping, data, geom, base_color, base_shape, base_size, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
