% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggscreeplot.R
\name{ggscreeplot}
\alias{ggscreeplot}
\title{Screeplot for Principal Components}
\usage{
ggscreeplot(pcobj, type = c("pev", "cev", "both"))
}
\arguments{
\item{pcobj}{an object returned by prcomp() or princomp()}

\item{type}{the type of scree plot.  'pev' corresponds proportion of explained variance, i.e. the eigenvalues divided by the trace. 'cev' corresponds to the cumulative proportion of explained variance, i.e. the partial sum of the first k eigenvalues divided by the trace.}
}
\description{
Screeplot for Principal Components
}
