% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggbiplot_geoms.R
\name{stat_centroid}
\alias{stat_centroid}
\title{Title}
\usage{
stat_centroid(mapping = NULL, data = NULL, geom = "text",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{typical mapping construct for ggplots. Typical in the form of "aes(x=, y=, color=, ...)"}

\item{data}{data to plot}

\item{geom}{how to visualize the data. "point" for points, "text" for labels}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function. In these plots this will mostly be "identity"}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. For creating a biplot this should be FALSE.}

\item{...}{other arguments passed on to layer.}
}
\description{
Title
}
