% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggbiplot_helper.R
\name{geom_selector}
\alias{geom_selector}
\title{Title}
\usage{
geom_selector(mapping, data, geom, base_color, base_shape, base_size, ...)
}
\arguments{
\item{mapping}{zie ggplot}

\item{data}{zie ggplot}

\item{geom}{welke geom wil je gebruiken: text | point | blank | line | path}

\item{base_color}{basiskleur grafiek}

\item{base_shape}{standaardsymbool}

\item{base_size}{standaard symboolgrootte}

\item{...}{andere parameters die gebruikt worden door ggplot geoms}
}
\description{
Title
}
