% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_ggvis_ggbiplot_helper.R
\name{get_display_data}
\alias{get_display_data}
\title{Title}
\usage{
get_display_data(x, data = NULL, choices = 1:2, display = "sites",
  scaling = "species", merge_by = "row.numbers", species_abbrev = FALSE,
  species_adjust = 1.1, remove_row_prefix = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
